Change project name from Alumirror to Alufo
Since it is likely that the left and right hand PCBs will be split, there won’t actually be much mirroring going on (also most split keyboards are mirrored left to right, so that not particularly special or unique!). Instead the name will emphasise the slight flying saucer look of the PCB with Al(uminium) + UFO = Alufo.