Fix boot access hole location
Because the lefthand side PCB will become the base for the righthand board (and vice versa), and because the boot button is no longer centred, the hole in the base to allow access to the button must be on the opposite side of the centreline to that button. This means the access hole needs to be a separate component and the boot switch footprint no longer needs a central hole.