Connect aluminium back to copper layer
Holes J5 and J6 will be filled with solder on the active PCB to connect the back aluminium layer (which is connected to the USB shield) with the mounting hole H3. H3 will in turn connect to the bottom layer PCB via the solder nut pressed against the exposed pad on the active PCB. The solder nut will connect to the screw and the screw will connect to the top plate. The upper middle plate will connect to the shield by virtue of being physically pressed to the active PCB’s back (if correctly oriented).