Add shield connectors
The USB ports will now be connected to the aluminium PCB core. This is done using a 0.6mm diameter wire. One end of the wire will be soldered to a pad on the PCB and the other will pass through a hole to friction fit against the aluminium core.

Also, R9, the 0Ω resistor on the lefthand VCC line has returned to allow a more direct ground route between the USB port and the MCU. However it is now a smaller 0805 size resistor.